SDR-332: refactor: comment 의 기존 기능을 command 와 query 패키지로 분리
