+++ be/src/main/java/com/jjikmuk/sikdorak/comment/api/CommentCommandApi.java
@@ -1,24 +1,19 @@
-package com.jjikmuk.sikdorak.comment.controller;
+package com.jjikmuk.sikdorak.comment.api;
 
 import static com.jjikmuk.sikdorak.common.ResponseCodeAndMessages.COMMENT_CREATE_SUCCESS;
 import static com.jjikmuk.sikdorak.common.ResponseCodeAndMessages.COMMENT_MODIFY_SUCCESS;
 import static com.jjikmuk.sikdorak.common.ResponseCodeAndMessages.COMMENT_REMOVE_SUCCESS;
-import static com.jjikmuk.sikdorak.common.ResponseCodeAndMessages.COMMENT_SEARCH_SUCCESS;
 
-import com.jjikmuk.sikdorak.comment.service.CommentService;
-import com.jjikmuk.sikdorak.comment.service.request.CommentCreateRequest;
-import com.jjikmuk.sikdorak.comment.service.request.CommentModifyRequest;
-import com.jjikmuk.sikdorak.comment.service.response.CommentSearchPagingResponse;
+import com.jjikmuk.sikdorak.comment.command.app.CommentService;
+import com.jjikmuk.sikdorak.comment.command.app.request.CommentCreateRequest;
+import com.jjikmuk.sikdorak.comment.command.app.request.CommentModifyRequest;
 import com.jjikmuk.sikdorak.common.aop.UserOnly;
-import com.jjikmuk.sikdorak.common.controller.CursorPageable;
-import com.jjikmuk.sikdorak.common.controller.request.CursorPageRequest;
 import com.jjikmuk.sikdorak.common.response.CommonResponseEntity;
 import com.jjikmuk.sikdorak.user.auth.controller.LoginUser;
 import com.jjikmuk.sikdorak.user.auth.domain.AuthenticatedUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -29,7 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/reviews/{reviewId}/comments")
 @RequiredArgsConstructor
-public class CommentController {
+public class CommentCommandApi {
 
 	private final CommentService commentService;
 
@@ -68,17 +63,5 @@ public class CommentController {
 		commentService.removeComment(reviewId, commentId, loginUser);
 
 		return new CommonResponseEntity<>(COMMENT_REMOVE_SUCCESS, HttpStatus.OK);
-	}
-
-	@GetMapping
-	public CommonResponseEntity<CommentSearchPagingResponse> searchComment(
-		@PathVariable long reviewId,
-		@AuthenticatedUser LoginUser loginUser,
-		@CursorPageable CursorPageRequest pagingRequest
-	) {
-		return new CommonResponseEntity<>(
-			COMMENT_SEARCH_SUCCESS,
-			commentService.searchCommentsByReviewIdWithPaging(reviewId, loginUser, pagingRequest),
-			HttpStatus.OK);
 	}
 }
