+++ be/src/main/java/com/jjikmuk/sikdorak/comment/command/domain/CommentRepository.java
@@ -1,6 +1,5 @@
-package com.jjikmuk.sikdorak.comment.repository;
+package com.jjikmuk.sikdorak.comment.command.domain;
 
-import com.jjikmuk.sikdorak.comment.domain.Comment;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
