+++ be/src/main/java/com/jjikmuk/sikdorak/comment/query/CommentDao.java
@@ -1,12 +1,9 @@
-package com.jjikmuk.sikdorak.comment.service;
+package com.jjikmuk.sikdorak.comment.query;
 
-import com.jjikmuk.sikdorak.comment.domain.Comment;
-import com.jjikmuk.sikdorak.comment.exception.NotFoundCommentException;
-import com.jjikmuk.sikdorak.comment.repository.CommentRepository;
-import com.jjikmuk.sikdorak.comment.service.request.CommentCreateRequest;
-import com.jjikmuk.sikdorak.comment.service.request.CommentModifyRequest;
-import com.jjikmuk.sikdorak.comment.service.response.CommentSearchPagingResponse;
-import com.jjikmuk.sikdorak.comment.service.response.CommentSearchResponse;
+import com.jjikmuk.sikdorak.comment.command.domain.Comment;
+import com.jjikmuk.sikdorak.comment.command.domain.CommentRepository;
+import com.jjikmuk.sikdorak.comment.query.response.CommentSearchPagingResponse;
+import com.jjikmuk.sikdorak.comment.query.response.CommentSearchResponse;
 import com.jjikmuk.sikdorak.common.controller.request.CursorPageRequest;
 import com.jjikmuk.sikdorak.common.controller.response.CursorPageResponse;
 import com.jjikmuk.sikdorak.common.exception.InvalidPageParameterException;
@@ -30,7 +27,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class CommentService {
+public class CommentDao {
 
     private static final long FIRST_CURSOR_ID = 0L;
     private static final long LAST_CURSOR_ID = 0L;
@@ -39,66 +36,6 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
     private final ReviewRepository reviewRepository;
-
-    @Transactional
-    public Comment createComment(long reviewId, LoginUser loginUser,
-        CommentCreateRequest commentCreateRequest) {
-
-        // 검증
-        User user = userRepository.findById(loginUser.getId())
-            .orElseThrow(NotFoundUserException::new);
-        Review review = reviewRepository.findById(reviewId)
-            .orElseThrow(NotFoundReviewException::new);
-
-        // 댓글 생성
-        Comment comment = new Comment(
-            review.getId(),
-            user.getId(),
-            commentCreateRequest.getContent());
-
-        return commentRepository.save(comment);
-    }
-
-    @Transactional
-    public void modifyComment(long reviewId, long commentId, LoginUser loginUser,
-        CommentModifyRequest modifyRequest) {
-        // 검증
-        validateReviewExists(reviewId);
-        User currentUser = userRepository.findById(loginUser.getId())
-            .orElseThrow(NotFoundUserException::new);
-        Comment comment = commentRepository.findById(commentId)
-            .orElseThrow(NotFoundCommentException::new);
-        validateModifiableUserOrThrow(comment, currentUser);
-
-        // 댓글 수정
-        comment.updateComment(modifyRequest.getContent());
-    }
-
-    private void validateReviewExists(long reviewId) {
-        if (!reviewRepository.existsById(reviewId)) {
-            throw new NotFoundReviewException();
-        }
-    }
-
-    private void validateModifiableUserOrThrow(Comment comment, User currentUser) {
-        if (!comment.isAuthor(currentUser.getId())) {
-            throw new UnauthorizedUserException();
-        }
-    }
-
-    @Transactional
-    public void removeComment(Long reviewId, Long commentId, LoginUser loginUser) {
-        // 검증
-        validateReviewExists(reviewId);
-        User currentUser = userRepository.findById(loginUser.getId())
-            .orElseThrow(NotFoundUserException::new);
-        Comment comment = commentRepository.findById(commentId)
-            .orElseThrow(NotFoundCommentException::new);
-        validateModifiableUserOrThrow(comment, currentUser);
-
-        // 댓글 삭제
-        comment.delete();
-    }
 
     @Transactional(readOnly = true)
     public CommentSearchPagingResponse searchCommentsByReviewIdWithPaging(long reviewId,
