+++ be/src/test/java/com/jjikmuk/sikdorak/documentationtest/comment/CommentCreateDocumentationTest.java
@@ -6,7 +6,7 @@ import static com.jjikmuk.sikdorak.documentationtest.comment.CommentSnippet.COMM
 import static io.restassured.RestAssured.given;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
-import com.jjikmuk.sikdorak.comment.service.request.CommentCreateRequest;
+import com.jjikmuk.sikdorak.comment.command.app.request.CommentCreateRequest;
 import com.jjikmuk.sikdorak.documentationtest.InitDocumentationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
