+++ be/src/test/java/com/jjikmuk/sikdorak/documentationtest/comment/CommentSnippet.java
@@ -11,8 +11,8 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWit
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 
-import com.jjikmuk.sikdorak.comment.service.request.CommentCreateRequest;
-import com.jjikmuk.sikdorak.comment.service.response.CommentSearchResponse;
+import com.jjikmuk.sikdorak.comment.command.app.request.CommentCreateRequest;
+import com.jjikmuk.sikdorak.comment.query.response.CommentSearchResponse;
 import com.jjikmuk.sikdorak.user.user.service.response.UserSimpleProfileResponse;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.restdocs.snippet.Snippet;
