+++ be/src/test/java/com/jjikmuk/sikdorak/integration/comment/CommentCreateIntegrationTest.java
@@ -3,9 +3,9 @@ package com.jjikmuk.sikdorak.integration.comment;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.jjikmuk.sikdorak.comment.domain.Comment;
-import com.jjikmuk.sikdorak.comment.service.CommentService;
-import com.jjikmuk.sikdorak.comment.service.request.CommentCreateRequest;
+import com.jjikmuk.sikdorak.comment.command.app.CommentService;
+import com.jjikmuk.sikdorak.comment.command.app.request.CommentCreateRequest;
+import com.jjikmuk.sikdorak.comment.command.domain.Comment;
 import com.jjikmuk.sikdorak.integration.InitIntegrationTest;
 import com.jjikmuk.sikdorak.review.exception.NotFoundReviewException;
 import com.jjikmuk.sikdorak.user.auth.controller.LoginUser;
