+++ be/src/test/java/com/jjikmuk/sikdorak/integration/comment/CommentRemoveIntegrationTest.java
@@ -3,10 +3,10 @@ package com.jjikmuk.sikdorak.integration.comment;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.jjikmuk.sikdorak.comment.domain.Comment;
+import com.jjikmuk.sikdorak.comment.command.app.CommentService;
+import com.jjikmuk.sikdorak.comment.command.domain.Comment;
+import com.jjikmuk.sikdorak.comment.command.domain.CommentRepository;
 import com.jjikmuk.sikdorak.comment.exception.NotFoundCommentException;
-import com.jjikmuk.sikdorak.comment.repository.CommentRepository;
-import com.jjikmuk.sikdorak.comment.service.CommentService;
 import com.jjikmuk.sikdorak.integration.InitIntegrationTest;
 import com.jjikmuk.sikdorak.review.domain.Review;
 import com.jjikmuk.sikdorak.review.exception.NotFoundReviewException;
